+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## 2024-12-08
+
+### Changed
+- Updated ElastiCache Serverless configuration to use `valkey` engine
+- Upgraded ElastiCache major engine version from `7` to `8`
+- Simplified KMS key import syntax from `aws_kms as kms` to `* as kms from 'aws-cdk-lib/aws-kms'`
+- Modified daily snapshot time from `"00:00-01:00"` to `"00:00"`
+
+### Added
+- Added TODO comment for future ElastiCache users and groups configuration
+
 ## 2024-12-07
 
 ### Added